requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
discord.py==1.6.0
idna==2.10
multidict==4.7.6
pyaml==20.4.0
PyYAML==5.3.1
termcolor==1.1.0
yarl==1.5.1
requests==2.25.1